requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-middleware==1.3.1
django-debug-toolbar==1.11
djangorestframework==3.9.4
idna==2.8
numpy==1.16.4
pandas==0.24.2
pdfkit==0.6.1
PyMySQL==0.9.3
pypinyin==0.35.3
python-dateutil==2.8.0
pytz==2019.1
qrcode==6.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
uWSGI==2.0.18
xlrd==1.2.0
xlwt==1.3.0
